+++ frontend/src/components/Norm/ArticleVersionWarning.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import WarningMessage from "~/components/Norm/WarningMessage.vue";
+import VersionWarningMessage from "~/components/Norm/VersionWarningMessage.vue";
 import type { Article } from "~/types";
 import { getStatusLabel } from "~/composables/useNormVersions";
 
@@ -17,10 +17,10 @@ const currentArticleStatus = computed(() =>
 </script>
 
 <template>
-  <WarningMessage
+  <VersionWarningMessage
     :current-version-status="currentArticleStatus"
     :in-force-version-link="inForceVersionLink"
     historical-warning-message="Paragraf einer historischen Fassung."
-    future-warning-message="Paragraph einer zukünftigen Fassung."
+    future-warning-message="Paragraf einer zukünftigen Fassung."
   />
 </template>
